chore: removed quill rich text editor, slate and tinymce providers
