+++ src/frontend/components/property-type/richtext/edit.tsx
@@ -1,47 +1,22 @@
-/* eslint-disable @typescript-eslint/no-use-before-define */
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-/* eslint-disable no-unused-expressions */
-import React, { FC, memo } from 'react'
-import {
-  FormGroup,
-  FormMessage,
-  RichText,
-  QuillOptions,
-  DefaultQuillToolbarOptions,
-} from '@adminjs/design-system'
-
+import { FormGroup, FormMessage, RichTextEditor } from '@adminjs/design-system'
+import React, { FC, memo, useCallback } from 'react'
 import { EditPropertyProps } from '../base-property-props'
 import { recordPropertyIsEqual } from '../record-property-is-equal'
 import { PropertyLabel } from '../utils/property-label'
 
-type CustomType = {
-  borderless?: boolean;
-  quill?: QuillOptions;
-}
-
 const Edit: FC<EditPropertyProps> = (props) => {
   const { property, record, onChange } = props
-  const value = record.params?.[property.path] ?? ''
+  const value = record.params?.[property.path] || ''
   const error = record.errors && record.errors[property.path]
 
-  const { props: propertyProps } = property
-
-  const { quill = {}, ...customProps } = propertyProps as CustomType || {}
-  quill.theme = quill.theme || 'snow'
-  quill.modules = {
-    toolbar: DefaultQuillToolbarOptions,
-    ...(quill.modules || {}),
-  }
+  const handleUpdate = useCallback((newValue: string) => {
+    onChange(property.path, newValue)
+  }, [])
 
   return (
     <FormGroup error={Boolean(error)}>
       <PropertyLabel property={property} />
-      <RichText
-        {...customProps}
-        value={value}
-        onChange={content => onChange(property.path, content)}
-        quill={quill}
-      />
+      <RichTextEditor value={value} onChange={handleUpdate} />
       <FormMessage>{error?.message}</FormMessage>
     </FormGroup>
   )
