+++ src/frontend/components/property-type/richtext/show.tsx
@@ -1,31 +1,25 @@
-import React, { ReactNode } from 'react'
-import { ValueGroup, Text, Box } from '@adminjs/design-system'
-
+import { Box, Text, ValueGroup } from '@adminjs/design-system'
+import React, { FC, useEffect, useRef } from 'react'
 import { EditPropertyProps } from '../base-property-props'
 
-export default class Show extends React.PureComponent<EditPropertyProps> {
-  private contentRef: React.RefObject<any>
+const Show: FC<EditPropertyProps> = (props) => {
+  const { property, record } = props
+  const contentRef = useRef<any>(null)
 
-  constructor(props: EditPropertyProps) {
-    super(props)
-    this.contentRef = React.createRef()
-  }
+  useEffect(() => {
+    if (contentRef.current) {
+      const value = record.params[property.path]
+      contentRef.current.innerHTML = value
+    }
+  }, [])
 
-  componentDidMount(): void {
-    const { property, record } = this.props
-    const value = record.params[property.path]
-    this.contentRef.current.innerHTML = value
-  }
-
-  render(): ReactNode {
-    const { property } = this.props
-
-    return (
-      <ValueGroup label={property.label}>
-        <Box variant="grey" border="default">
-          <Text ref={this.contentRef} />
-        </Box>
-      </ValueGroup>
-    )
-  }
+  return (
+    <ValueGroup label={property.label}>
+      <Box py="xl" px={['0', 'xl']} border="default">
+        <Text ref={contentRef} />
+      </Box>
+    </ValueGroup>
+  )
 }
+
+export default Show
