--- src/frontend/components/property-type/slate/edit.tsx
@@ -1,20 +1,0 @@
-import { FormGroup, FormMessage } from '@adminjs/design-system'
-import React, { FC, memo } from 'react'
-import { EditPropertyProps } from '../base-property-props'
-import { recordPropertyIsEqual } from '../record-property-is-equal'
-import { PropertyLabel } from '../utils/property-label'
-
-const Edit: FC<EditPropertyProps> = (props) => {
-  const { property, record, onChange } = props
-  const value = record.params?.[property.path] || ''
-  const error = record.errors && record.errors[property.path]
-
-  return (
-    <FormGroup error={Boolean(error)}>
-      <PropertyLabel property={property} />
-      <FormMessage>{error?.message}</FormMessage>
-    </FormGroup>
-  )
-}
-
-export default memo(Edit, recordPropertyIsEqual)
