--- src/frontend/components/property-type/slate/index.ts
@@ -1,9 +1,0 @@
-import Edit from './edit'
-import Show from './show'
-import List from './list'
-
-export {
-  Edit as edit,
-  Show as show,
-  List as list,
-}
