--- src/frontend/components/property-type/slate/show.tsx
@@ -1,31 +1,0 @@
-import React, { ReactNode } from 'react'
-import { ValueGroup, Text, Box } from '@adminjs/design-system'
-
-import { EditPropertyProps } from '../base-property-props'
-
-export default class Show extends React.PureComponent<EditPropertyProps> {
-  private contentRef: React.RefObject<any>
-
-  constructor(props: EditPropertyProps) {
-    super(props)
-    this.contentRef = React.createRef()
-  }
-
-  componentDidMount(): void {
-    const { property, record } = this.props
-    const value = record.params[property.path]
-    this.contentRef.current.innerHTML = value
-  }
-
-  render(): ReactNode {
-    const { property } = this.props
-
-    return (
-      <ValueGroup label={property.label}>
-        <Box py="xl" px={['0', 'xl']} border="default">
-          <Text ref={this.contentRef} />
-        </Box>
-      </ValueGroup>
-    )
-  }
-}
