--- src/frontend/components/property-type/tinymce/edit.tsx
@@ -1,27 +1,0 @@
-import { FormGroup, FormMessage } from '@adminjs/design-system'
-import { Editor } from '@tinymce/tinymce-react'
-import React, { FC, memo, useCallback, useRef } from 'react'
-import { EditPropertyProps } from '../base-property-props'
-import { recordPropertyIsEqual } from '../record-property-is-equal'
-import { PropertyLabel } from '../utils/property-label'
-
-const Edit: FC<EditPropertyProps> = (props) => {
-  const { property, record, onChange } = props
-  const value = record.params?.[property.path] || ''
-  const error = record.errors && record.errors[property.path]
-  const editorRef = useRef(null)
-
-  const handleUpdate = useCallback((newValue: string) => {
-    onChange(property.path, newValue)
-  }, [])
-
-  return (
-    <FormGroup error={Boolean(error)}>
-      <PropertyLabel property={property} />
-      <Editor onChange={console.log} initialValue={value} />
-      <FormMessage>{error?.message}</FormMessage>
-    </FormGroup>
-  )
-}
-
-export default memo(Edit, recordPropertyIsEqual)
